pyproject.toml
[project]
name = "haystackneo4jrecomend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.2.0",
    "haystack>=0.42",
    "haystack-ai>=2.21.0",
    "neo4j>=6.0.3",
    "neo4j-haystack>=2.2.1",
    "ollama-haystack>=5.3.0",
    "python-dotenv>=1.2.1",
]
